Add temporary fix to auth state on expired tokens

+++ services/authenticator.ts
@@ -1,21 +1,30 @@
 import { AuthUser } from "declarations";
+import { FirebaseError } from "firebase-admin";
 import { GetServerSidePropsContext, NextApiRequest, NextApiResponse } from "next";
 import { destroyCookie } from "nookies";
 import { authAdmin } from "services/firebaseAdmin";
 
 export const getAuthUser: (
   context: GetServerSidePropsContext | { req: NextApiRequest; res: NextApiResponse }
-) => Promise<AuthUser> = async (context) => {
+) => Promise<AuthUser | false> = async (context) => {
   try {
-    const decodedToken = await authAdmin.verifyIdToken(String(context.req.cookies.authToken));
-    return {
-      displayName: decodedToken.name || null,
-      email: decodedToken.email || null,
-      isAnonymous: decodedToken.firebase.sign_in_provider === "anonymous",
-      photoURL: decodedToken.picture || null,
-      uid: decodedToken.uid,
-    };
+    if (context.req.cookies.authToken) {
+      const decodedToken = await authAdmin.verifyIdToken(context.req.cookies.authToken);
+      return {
+        displayName: decodedToken.name || null,
+        email: decodedToken.email || null,
+        isAnonymous: decodedToken.firebase.sign_in_provider === "anonymous",
+        photoURL: decodedToken.picture || null,
+        uid: decodedToken.uid,
+      };
+    } else {
+      return null;
+    }
   } catch (err) {
+    const firebaseError = err as FirebaseError;
+    if (firebaseError.code === "auth/argument-error") {
+      return false;
+    }
     destroyCookie(context, "authToken", {
       path: "/",
     });
